Copying image buffers

+++ src/gfx/mod.rs
@@ -1,8 +1,6 @@
-use ::math::size::Size;
-
 pub mod image;
 pub mod screen;
 
 const SCREEN_TOTAL_PIXELS: isize = 160 * 144;
 
-pub type Color = [i32; 4];
+pub type Color = [u8; 4];
